requirements.txt
# YMERA Project Agent - Production Dependencies
# Python 3.11+ required

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Caching & Queue
redis[hiredis]==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
pyjwt==2.8.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# File Storage
aiofiles==23.2.1

# Monitoring & Observability
prometheus-client==0.19.0

# OpenTelemetry (Optional - for distributed tracing)
# Uncomment to enable:
# opentelemetry-api==1.21.0
# opentelemetry-sdk==1.21.0
# opentelemetry-instrumentation-fastapi==0.42b0
# opentelemetry-instrumentation-sqlalchemy==0.42b0
# opentelemetry-instrumentation-redis==0.42b0
# opentelemetry-exporter-jaeger==1.21.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
pyyaml==6.0.1
pytz==2023.3
python-dateutil==2.8.2
structlog==23.2.0
psutil==5.9.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==21.0.0
bandit==1.7.5
